Combining the region, test, and module generators into a single file - holidays/definitions/generator.rb. The module/test generators were already refactored in a previous revision, but the regions generator needed to be converted from a class "`call()`" style thing.
